+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- Add shared hashset.
 - Add entity observer manager, which is used to send components only for entities that are observed by - client.
 - Add partial (de)serialization of component storages.
 - Add generator manager.